Fix age group null values
