+++ lib/ui/screens/statistics/components/statistics_vertical.dart
@@ -11,7 +11,6 @@
 ///    You should have received a copy of the GNU General Public License
 ///    along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-import 'package:covid19mobile/resources/style/text_styles.dart';
 import 'package:covid19mobile/ui/assets/colors.dart';
 import 'package:covid19mobile/ui/screens/statistics/components/number_and_percentage.dart';
 import 'package:covid19mobile/ui/screens/statistics/components/see_details.dart';
