+++ lib/ui/screens/statistics/model/age_group_by_sex.dart
@@ -1,0 +1,31 @@
+///    This program is free software: you can redistribute it and/or modify
+///    it under the terms of the GNU General Public License as published by
+///    the Free Software Foundation, either version 3 of the License, or
+///    (at your option) any later version.
+///
+///    This program is distributed in the hope that it will be useful,
+///    but WITHOUT ANY WARRANTY; without even the implied warranty of
+///    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+///    GNU General Public License for more details.
+///
+///    You should have received a copy of the GNU General Public License
+///    along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+import 'package:flutter/material.dart';
+
+class AgeGroupBySex {
+  final int order;
+  final String ageGroup;
+  double male;
+  double female;
+
+  AgeGroupBySex({
+    @required this.order,
+    @required this.ageGroup,
+    male,
+    female,
+  }) {
+    this.male = male == null ? 0 : male;
+    this.female = female == null ? 0 : female;
+  }
+}
